Finished PartService tests Added a numeric id to the brand shifted entities into their own file Renamed NewProductRequest to NewPartRequest Implemented validation on the NewPartRequest Added an error advice Added various methods to the part repo interface to remove filtering logic from the service layer Did consistency naming clean-up and replace production year from long to normal int for model.

+++ src/main/java/dev/ioannis/anemosparts/entity/ModelEntityId.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ModelEntityId {
     private String name;
-    private Long productionYear;
+    private Integer productionYear;
 }
